Move delegate stats checks out of API tests
Commit supported by Codex/GPT5.5 and Windsurf/Opus4.7
